cleanup will not delete the run document or folder if it didn't send to the server or it was unsuccessful.
git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/sierra-tool@817 d02ae147-ee12-4a35-9452-91c89e8b2592

+++ src/com/surelogic/sierra/tool/ant/SierraAnalysis.java
@@ -355,7 +355,11 @@ public class SierraAnalysis extends Task {
 	private void cleanup() {
 		log("Cleaning up...", org.apache.tools.ant.Project.MSG_INFO);
 		tools.cleanup();
-		tmpFolder.delete();
+		if(serverURL != null){
+    		// FIXME should not delete rundocument if it didn't send to the server or the upload was unsuccessful
+			runDocument.delete();
+    		tmpFolder.delete();
+		}
 	}
 
 	/**
